requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.3.4
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.6
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.4
django-allauth==0.49.0
django-configurations==2.3.1
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-extensions==3.1.3
django-filter==2.4.0
django-jalali==5.0.0
django-otp==1.1.3
django-registration==3.2
django-taggit==1.4.0
django-versatileimagefield==2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
httpie==3.1.0
idna==2.10
inflection==0.5.1
ipython==7.24.1
ipython-genutils==0.2.0
itypes==1.2.0
jdatetime==3.8.0
jedi==0.18.0
Jinja2==3.1.1
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.2
multidict==6.0.2
mysqlclient==2.0.3
oauthlib==3.1.1
packaging==21.3
parso==0.8.2
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.24
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==3.0.7
PySocks==1.7.1
python-magic==0.4.25
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
traitlets==5.0.5
uritemplate==4.1.1
urllib3==1.26.5
wcwidth==0.2.5
